(PE-36953) Add reusable workflow for exporting GH issues to JIRA
This commit defines a reusable workflow to export a GH issue to JIRA in response
to a `triaged` label being applied to the issue.

The workflow is idempotent, so that if `triaged` label is removed and readded,
then we won't create another JIRA issue. Idempotency is implemented by checking
to see if the `github-actions[bot]` already added a comment with specific text.

Otherwise, the `github-actions[bot]` uses the `jira-api-token` secret to log
into JIRA and create an issue. The `jira-base-url`, `jira-user-email` and
`jira-project` variables must also be specified. The action adds a comment to
the GH issue with the URL of the newly created JIRA ticket.

    Migrated issue to [FACT-XXX]($JIRA_BASE_URL/browse/FACT-XXX)

The newly created JIRA ticket includes a reference back to the originating GH
issue:

    Originally reported in https://github.com/puppetlabs/<project>/issues/<id>

This workflow should be called from another workflow, for example, for facter:

    on:
      issues:
        types: [labeled]

    permissions:
      issues: write

    jobs:
      export:
        uses: "puppetlabs/phoenix-github-actions/.github/workflows/jira.yml@main"
        with:
          jira-project: FACT
          jira-base-url: ${{ vars.JIRA_BASE_URL }}
          jira-user-email: ${{ vars.JIRA_USER_EMAIL }}
        secrets:
          jira-api-token: ${{ secrets.JIRA_ISSUES_ACTION }}

The outer workflow must declare that it requires `write` permission to issues so
that the reusable workflow can add a comment.

Only users with `triage` github permission are allowed to apply (and dismiss) labels to
issues, so we don't need to worry about random users creating JIRA tickets.