Use pandoc to convert GFM to JIRA wiki format
Although the JIRA Cloud UI accepts Github Flavored Markdown (GFM), it
autoconverts to Atlassian Document Format (ADF). The same functionality is not
available in JIRA's v2 or v3 REST API. So any markdown in a github issue did not
render correctly when displayed in JIRA.

This commit uses pandoc to convert from GFM to JIRA formats.

Since the github issue description may contain unsafe shell characters, we use
actions to read & write the description. This avoids having to read file
contents into an environment variable and trying to pass the multiline value to
the next step[1].

[1] https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#example-of-a-multiline-string