(PA-5872) Create reusable workflow to import issues
Creates a reusable workflow to import a single issue from Jira to GitHub.

The caller workflow must pass 4 inputs:

    jira-key
    jira-base-url
    jira-user-email
    jira-api-token

The first input is the issue to import, e.g. PUP-123. The other 3 inputs are the
same as what's used in the workflow that exports issues from GitHub to Jira.

This workflow executes a ruby script that relies on the `jira-ruby` gem,
`pandoc` executable (to convert Jira wiki to GFM) and the `gh` CLI. The latter
is "preinstalled on all GitHub-hosted runners"[1]

[1] https://docs.github.com/en/actions/using-workflows/using-github-cli-in-workflows