(PA-5940) Allow Jira issues to be created from pull requests
This commit makes it possible to apply a label to a pull request and have it
create a JIRA ticket, just like it does for issues.. In order to use this, the
calling workflow must specify (note the differences in underscore vs dash):

    on:
     pull_requests:
       types: [labeled]
    permissions:
      pull-requests: write

When a label is applied to an issue or pull request, GitHub triggers the
workflow with a `labeled` event and either a github.event.issue[1] or
github.event.pull_request[2] payload, respectively. So this PR checks for either
the issue or pull_request as the payload.

If the workflow is trigged due to a pull request, then
`github.event.issue.number` will safely evaluate to null. In other words, we
don't need to write `(github.event.issue && github.event.issue.number) || ...`

If a PR is submitted with an empty body, then provide a default message,
otherwise the `write-file-action` will error with:

    Error: Input required and not supplied: contents

[1] https://docs.github.com/en/webhooks/webhook-events-and-payloads?actionType=labeled#issue
[2] https://docs.github.com/en/webhooks/webhook-events-and-payloads?actionType=labeled#pull_request