(maint) Update windows action runner image to windows-2022
The Windows 2019 image will be removed on 2025-06-30 and is currently in a
brownout period. For more details, see
https://github.com/actions/runner-images/issues/12045.

Due to this, this commit updates the unit test workflows to use the
windows-2022 runner image instead of windows-2019. windows-2022 was chosen
instead of windows-2025 because it's a supported platform for both Puppet 7
and 8.